+++ components/clientComponents/forms/SaveAndResume/MobileDrawer.tsx
@@ -3,79 +3,31 @@ import { SaveProgressIcon, UploadIcon } from "@serverComponents/icons";
 import { Button } from "@clientComponents/globals";
 import Link from "next/link";
 import { Language } from "@lib/types/form-builder-types";
-import { useGCFormsContext } from "@lib/hooks/useGCFormContext";
-import { FormValues } from "@lib/formContext";
-import { getReviewItems } from "../Review/helpers";
-import { slugify } from "@lib/client/clientHelpers";
-import { getStartLabels } from "@lib/utils/form-builder/i18nHelpers";
 import { generateDownloadHtml } from "@lib/saveAndResume/actions";
 import { downloadDataAsBlob } from "@lib/downloadDataAsBlob";
 import { useTranslation } from "@i18n/client";
+import { useFormSubmissionData } from "@lib/hooks/useFormSubmissionData";
 
 export const MobileDrawer = ({
   drawerOpen,
   setDrawerOpen,
   formId,
   language,
-  formTitleEn,
-  formTitleFr,
   type,
 }: {
   drawerOpen: boolean;
   setDrawerOpen: React.Dispatch<React.SetStateAction<boolean>>;
   formId: string;
   language: Language;
-  formTitleEn: string;
-  formTitleFr: string;
   type: "confirm" | "progress";
 }) => {
   const { t } = useTranslation("common");
 
-  const {
-    groups,
-    getValues,
-    formRecord,
-    getGroupHistory,
-    matchedIds,
-    getProgressData,
-    submissionId,
-    submissionDate,
-  } = useGCFormsContext();
-
-  const formValues: void | FormValues = getValues();
-  const groupHistoryIds = getGroupHistory();
-  if (!formValues || !groups) throw new Error("Form values or groups are missing");
-
-  const reviewItems = getReviewItems({
-    formElements: formRecord.form.elements,
-    formValues,
-    groups,
-    groupHistoryIds,
-    matchedIds,
-    language,
-  });
+  const { fileName, getOptions } = useFormSubmissionData({ language, type });
 
   const handleSave = async () => {
-    const title = language === "en" ? formTitleEn : formTitleFr;
-
-    const fileName = `${slugify(title)}-${formId}.html`;
-
-    const options = {
-      formTitle: title,
-      type,
-      formId,
-      reviewItems,
-      formResponse: Buffer.from(JSON.stringify(getProgressData()), "utf8").toString("base64"),
-      language,
-      startSectionTitle: getStartLabels()[language],
-      submissionId,
-      submissionDate,
-    };
-
-    const html = await generateDownloadHtml(options);
-
+    const html = await generateDownloadHtml(getOptions());
     await downloadDataAsBlob(html.data, fileName, { "text/html": [".html"] });
-
     setDrawerOpen(false);
   };
